backend/requirements.txt
alembic==1.8.1
apiclient==1.0.4
fastapi==0.92.0
pydantic[email]==1.10.2
uvicorn==0.20.0
sqlmodel==0.0.8
PyJWT==2.6.0
psycopg2-binary==2.9.5
requests==2.28.2
python-dateutil==2.8.2
celery[redis]==5.2
